requirements.txt
mock
freezegun
keboola.component==1.4.0
keboola.http-client
dateparser
regex
pandas==1.2.2